[Rename] Diary에서 Record로 이름 변경
