+++ Frontend/src/components/RecordHeader.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { View, Text } from "react-native";
-import styles from "../styles/diaryHeaderStyles";
+import styles from "../styles/recordHeaderStyles";
 
-export default function DiaryHeader() {
+export default function RecordHeader() {
     const today = new Date();
     const dateStr = `${today.getMonth() + 1}월 ${today.getDate()}일 ${
         ["일", "월", "화", "수", "목", "금", "토"][today.getDay()]
@@ -11,7 +11,7 @@ export default function DiaryHeader() {
     return (
         <View style={styles.header}>
             <Text style={styles.dateOnly}>{dateStr}</Text>
-            <Text style={styles.titleText}>오늘의 일기</Text>
+            <Text style={styles.titleText}>오늘의 기록</Text>
         </View>
     );
 }
