+++ Frontend/src/screens/Home.tsx
@@ -56,7 +56,7 @@ export default function Home() {
                         icon="book-open-variant"
                         title="일기"
                         subtitle="오늘 하루, 나의 마음 기록하기"
-                        onPress={() => navigation.navigate("Diary")}
+                        onPress={() => navigation.navigate("Record")}
                     />
                     <HomeButton
                         icon="target"
