+++ Frontend/src/styles/recordHeaderStyles.ts
