+++ Frontend/src/styles/recordStyles.ts
