+++ Frontend/src/styles/submitRecordButtonStyles.ts
